+++ RELEASE_NOTES.md
@@ -14,3 +14,4 @@
 ## New Features
 
 - The client now inherits from `frequenz.client.base.BaseApiClient`, so it provides a few new features, like `disconnect()`ing or using it as a context manager. Please refer to the [`BaseApiClient` documentation](https://frequenz-floss.github.io/frequenz-client-base-python/latest/reference/frequenz/client/base/client/#frequenz.client.base.client.BaseApiClient) for more information on these features.
+- The client now supports setting reactive power for components through the new `set_reactive_power` method.